persist files

+++ Program.cs
@@ -38,25 +38,29 @@ static async Task StartAnalysisAsync(IHost host)
                             item.Link + item.Title);
         Console.WriteLine(md5);
 
-        var result = await item.Link.GetStringAsync();
-        var endIndex = result.IndexOf("</h2>", StringComparison.Ordinal);
-        var startIndex = result.Substring(0, endIndex).LastIndexOf('>');
-        var title = result.Substring(startIndex + 1, endIndex - startIndex - 1);
-
         Directory.CreateDirectory("./MD5");
         Directory.CreateDirectory("./Title");
 
-        var md5FullPath = Path.Combine("./MD5", CreateMD5(item.Id));
-        await File.WriteAllTextAsync(
-            md5FullPath,
-            md5,
-            tokenSource.Token);
+        var id = item.Id.Substring(0, item.Id.LastIndexOf('-'));
+        var md5FullPath = Path.Combine("./MD5", CreateMD5(id));
+        if (!File.Exists(md5FullPath) || File.ReadAllText(md5FullPath) != md5)
+        {
+            var result = await item.Link.GetStringAsync();
+            var endIndex = result.IndexOf("</h2>", StringComparison.Ordinal);
+            var startIndex = result.Substring(0, endIndex).LastIndexOf('>');
+            var title = result.Substring(startIndex + 1, endIndex - startIndex - 1);
 
-        var titleFullPath = Path.Combine("./Title", CreateMD5(item.Id));
-        await File.WriteAllTextAsync(
-            titleFullPath,
-            title,
-            tokenSource.Token);
+            await File.WriteAllTextAsync(
+                md5FullPath,
+                md5,
+                tokenSource.Token);
+
+            var titleFullPath = Path.Combine("./Title", CreateMD5(id));
+            await File.WriteAllTextAsync(
+                titleFullPath,
+                title,
+                tokenSource.Token);
+        }
     }
 
     Environment.Exit(0);
